fix: add DBTests
