+++ .github/workflows/swift-built-test.yaml
@@ -17,4 +17,4 @@ jobs:
     - name: Build
       run: swift build -v
     - name: Run tests
-      run: swift test -v
+      run: swift test -v --skip DBTests
